+++ pytorch_lightning/loggers/mlflow.py
@@ -17,8 +17,11 @@ MLflow Logger
 """
 import re
 from argparse import Namespace
+from pathlib import Path
 from time import time
 from typing import Any, Dict, Optional, Union
+
+import matplotlib.pyplot as plt
 
 from pytorch_lightning import _logger as log
 from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_experiment
@@ -78,6 +81,7 @@ class MLFlowLogger(LightningLoggerBase):
             Defaults to `./mlflow` if `tracking_uri` is not provided.
             Has no effect if `tracking_uri` is provided.
         prefix: A string to put at the beginning of metric keys.
+        figure_file_extension: File extension with which matplotlib saves figure
 
     Raises:
         ImportError:
@@ -93,6 +97,7 @@ class MLFlowLogger(LightningLoggerBase):
         tags: Optional[Dict[str, Any]] = None,
         save_dir: Optional[str] = './mlruns',
         prefix: str = '',
+        figure_file_extension='.png',
     ):
         if mlflow is None:
             raise ImportError(
@@ -110,6 +115,7 @@ class MLFlowLogger(LightningLoggerBase):
         self.tags = tags
         self._prefix = prefix
         self._mlflow_client = MlflowClient(tracking_uri)
+        self._figure_file_extension = figure_file_extension
 
     @property
     @rank_zero_experiment
@@ -182,6 +188,18 @@ class MLFlowLogger(LightningLoggerBase):
                 k = new_k
 
             self.experiment.log_metric(self.run_id, k, v, timestamp_ms, step)
+
+    @rank_zero_only
+    def log_figure(self, name: str, figure: plt.figure, step: Optional[int] = None, close: bool = True) -> None:
+
+        filename = self.save_dir + '/' + name + f"_step_{step}" + self._figure_file_extension
+        figure.savefig(filename)
+        self.experiment.log_artifact(self.run_id, filename, artifact_path="figure_" + name)
+
+        Path(filename).unlink(missing_ok=False)  # delete temporary file
+
+        if close:
+            plt.close(figure)
 
     @rank_zero_only
     def finalize(self, status: str = 'FINISHED') -> None: