+++ pytorch_lightning/loggers/tensorboard.py
@@ -20,6 +20,7 @@ import os
 from argparse import Namespace
 from typing import Any, Dict, Optional, Union
 
+import matplotlib.pyplot as plt
 import torch
 from torch.utils.tensorboard import SummaryWriter
 from torch.utils.tensorboard.summary import hparams
@@ -203,6 +204,10 @@ class TensorBoardLogger(LightningLoggerBase):
                 except Exception as ex:
                     m = f'\n you tried to log {v} which is not currently supported. Try a dict or a scalar/tensor.'
                     type(ex)(ex.message + m)
+
+    @rank_zero_only
+    def log_figure(self, name: str, figure: plt.figure, step: Optional[int] = None, close: bool = True) -> None:
+        self.experiment.add_figure(tag=name, figure=figure, global_step=step, close=close)
 
     @rank_zero_only
     def log_graph(self, model: LightningModule, input_array=None):