+++ pytorch_lightning/loggers/wandb.py
@@ -19,6 +19,7 @@ import os
 from argparse import Namespace
 from typing import Any, Dict, Optional, Union
 
+import matplotlib.pyplot as plt
 import torch.nn as nn
 
 from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_experiment
@@ -199,6 +200,13 @@ class WandbLogger(LightningLoggerBase):
             self.experiment.log({**metrics, 'trainer_step': (step + self._step_offset)})
         else:
             self.experiment.log(metrics)
+
+    @rank_zero_only
+    def log_figure(self, name: str, figure: plt.figure, step: Optional[int] = None, close: bool = True) -> None:
+        self.experiment.log({name: wandb.Image(figure)}, step=step)
+
+        if close:
+            plt.close(figure)
 
     @property
     def save_dir(self) -> Optional[str]: